backend/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-qdrant
qdrant-client
sentence-transformers
openai>=1.55.0
pypdf==3.17.4
python-dotenv==1.0.0
rank-bm25
fastapi
uvicorn
pydantic
